Refactor: dualPane 여부를 UseCase를 통해서 받아오도록 변경
