+++ app/src/main/java/com/example/ifkakao/presentation/home/HomeViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.ifkakao.presentation
+package com.example.ifkakao.presentation.home
 
 import androidx.lifecycle.ViewModel
 import com.example.ifkakao.domain.use_case.GetIsDualPaneUseCase
@@ -6,10 +6,8 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import javax.inject.Inject
 
 @HiltViewModel
-class MainViewModel @Inject constructor(
+class HomeViewModel @Inject constructor(
     private val getIsDualPaneUseCase: GetIsDualPaneUseCase
 ) : ViewModel() {
-    var backupStatusBarColor: Int? = null
-
     fun getIsDualPane() = getIsDualPaneUseCase()
 }
